Use window *'s instead of layout *'s in layout api for simplicity

+++ lib/winmgr/splitter.c
@@ -31,7 +31,7 @@ void splitter_paint(splitter *spltr)
     }
 }
 
-uint32_t splitter_proc(splitter *spltr, int id, const message_data *data)
+uintptr_t splitter_proc(splitter *spltr, int id, const message_data *data)
 {
     switch (id)
     {
